execution/stagedsync: reclaim the displaced credit, pin it to its incarnation
recordWorkerWrites dropped the fee-merge product from feeMergeTemp without
reclaiming it, so every re-execution of an already-credited tx sent the merged
set's pooled maps to GC.

feeMergeTemp now carries the version each credit was computed for, so a stale
credit cannot be handed to calcFees even if it outlived the recorded set it
lived in — the CollectorWrites entries an earlier round wrote belong to that
same incarnation.

recordFeeMerge does the MergeInto and the RecordWrites itself, which keeps the
merge product and the tx's recorded write set from drifting apart and makes the
"a temp is never an execResult's TxOut" ownership rule local to the one function
that populates the map.

The fee-credit tests move into exec3_finalize_test.go and
exec3_fee_merge_temp_test.go, next to the fixtures and the blockExecutor
white-box tests they belong with.
